+++ services/svc-api-gateway/internal/tools/generate.go
@@ -1,4 +1,4 @@
 package tools
 
-//go:generate go tool github.com/oapi-codegen/oapi-codegen/v2/cmd/oapi-codegen --config=../../../../build/oapi/codegen-public.yaml ../../../../docs/contracts/openapi/devices/v1/public/specs-swagger.json
-//go:generate go tool github.com/oapi-codegen/oapi-codegen/v2/cmd/oapi-codegen --config=../../../../build/oapi/codegen-admin.yaml ../../../../docs/contracts/openapi/devices/v1/public/specs-swagger.json
+//go:generate go tool github.com/oapi-codegen/oapi-codegen/v2/cmd/oapi-codegen --config=../../../../build/oapi/codegen-public.yaml ../../../../docs/contracts/openapi/devices/v1/public/swagger-pact.json
+//go:generate go tool github.com/oapi-codegen/oapi-codegen/v2/cmd/oapi-codegen --config=../../../../build/oapi/codegen-admin.yaml ../../../../docs/contracts/openapi/devices/v1/public/swagger-pact.json